Implements a setup procedure for the hangman game.
The `Game` struct in question pre-loads all hangman `frames`, selects a random
`word` from words.txt and initializes the `life` points to `frames.size()`. It
also explodes the word into a vector of `chars` that will be later used in the
game logic.

For this purpose, a few helper methods have been added to `game.cpp`:

- random_range
- random_choice
- split_string